refactor(cli): move the four inline test modules to sibling files (#1918)
First crate of the per-crate migration campaign. Pure move: each
#[cfg(test)] mod tests block becomes a <stem>_tests.rs sibling declared
via #[path] (the import.rs pattern the guard prescribes), content
unchanged except rustfmt re-wrapping lines that lost one indent level.
Test count identical before/after (222 on the main target). Baseline
shrinks by the four velesdb-cli lines (697 lines of frozen debt paid).
